+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file. This projec
 ## [Unreleased]
 
 ### Added
+- Choice where to place markup: head or body section (#21)
 - API for retrieving lists of types (#19)
 
 ## [1.2.0] - 2019-09-03